% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_dropouts.R
\docType{data}
\name{ca_dropouts}
\alias{ca_dropouts}
\title{School Dropouts by Race & Gender}
\format{
\tabular{ll}{
\bold{Field.Name} \tab \bold{Description} \cr
CDSCode    \tab This 14-digit code is the official, unique identification of a school within California. The first two digits identify the county, the next five digits identify the school district, and the last seven digits identify the school.                                                  \cr
\tab Please note that a CDS code ending in '0000000' indicates a district record not a school.                                                                                                                                                                                             \cr
NCESDist   \tab This field represents the 7-digit National Center for Educational Statistics (NCES) school district identification number. The first 2 digits identify the state and the last 5 digits identify the school district. Combined, they make a unique 7-digit ID for each school district.\cr
NCESSchool \tab This field represents the 5-digit NCES school identification number. The NCESSchool combined with the NCESDist form a unique 12-digit ID for each school.                                                                                                                             \cr
StatusType \tab This field identifies the status of the district. Definitions of the valid status types are listed below:                                                                                                                                                                             \cr
\tab Active: The district is in operation and providing instructional services.                                                                                                                                                                                                            \cr
\tab Closed: The district is not in operation and no longer providing instructional services.                                                                                                                                                                                              \cr
\tab Merged: The district has combined with another district or districts.                                                                                                                                                                                                                 \cr
\tab Pending: The district has not opened for operation and instructional services yet, but plans to open within the next 9<d0>12 months.
}
}
\source{
California Department of Education Data Reporting Office

This data was downloaded 2023/01/20 at 13:00:29 EST and gathered from:
\url{https://www.cde.ca.gov/ds/ad/filesdropouts.asp}.
}
\usage{
ca_dropouts
}
\description{
Historical data for grade seven through twelve dropouts and enrollment by
race/ethnic designation and gender by school.
}
\details{
\tabular{lll}{
   \strong{Field Name} \tab  \tab \strong{Description} \cr
   CDS_CODE \tab \verb{  } \tab This 14-digit code is the official, unique identification of a  school within California. The first two digits identify the county, the  next five digits identify the school district, and the last seven digits  identify the school. \cr
   ETHNIC \tab \verb{  } \tab This is a coded field for racial/ethnic designation. The racial/ethnic designations are coded as follows: \cr
    \tab \verb{  } \tab ∙ Code 0 = Not reported \cr
    \tab \verb{  } \tab ∙ Code 1 = American Indian or Alaska Native, Not Hispanic \cr
    \tab \verb{  } \tab ∙ Code 2 = Asian, Not Hispanic \cr
    \tab \verb{  } \tab ∙ Code 3 = Pacific Islander, Not Hispanic \cr
    \tab \verb{  } \tab ∙ Code 4 = Filipino, Not Hispanic \cr
    \tab \verb{  } \tab ∙ Code 5 = Hispanic or Latino \cr
    \tab \verb{  } \tab ∙ Code 6 = African American, not Hispanic (formerly known as Black, not Hispanic) \cr
    \tab \verb{  } \tab ∙ Code 7 = White, not Hispanic \cr
    \tab \verb{  } \tab ∙ Code 9 = Two or More Races, Not Hispanic. \cr
   GENDER \tab \verb{  } \tab This field is a coded field identifying gender. The gender is coded as follows: \cr
    \tab \verb{  } \tab ∙ M = Male \cr
    \tab \verb{  } \tab ∙ F = Female \cr
   E7 \tab \verb{  } \tab Enrollment in grade seven. (Not included in enrollment total (ETOT) field below). \cr
   E8 \tab \verb{  } \tab Enrollment in grade eight. (Not included in enrollment total (ETOT) field below). \cr
   E9 \tab \verb{  } \tab Enrollment in grade nine. \cr
   E10 \tab \verb{  } \tab Enrollment in grade ten. \cr
   E11 \tab \verb{  } \tab Enrollment in grade eleven. \cr
   E12 \tab \verb{  } \tab Enrollment in grade twelve. \cr
   EUS \tab \verb{  } \tab Enrollment in ungraded secondary classes. \cr
   ETOT \tab \verb{  } \tab Total enrollment for grades nine through twelve and ungraded secondary. (Does not include grades seven and eight). \cr
   D7 \tab \verb{  } \tab Dropouts in grade seven. (Not included in dropout total (DTOT) field below) \cr
   D8 \tab \verb{  } \tab Dropouts in grade eight. (Not included in dropout total (DTOT) field below). \cr
   D9 \tab \verb{  } \tab Dropouts in grade nine. \cr
   D10 \tab \verb{  } \tab Dropouts in grade ten. \cr
   D11 \tab \verb{  } \tab Dropouts in grade eleven. \cr
   D12 \tab \verb{  } \tab Dropouts in grade twelve. \cr
   DUS \tab \verb{  } \tab Dropouts in ungraded secondary classes. \cr
   DTOT \tab \verb{  } \tab Total dropouts for grades nine through twelve and ungraded secondary. (Does not include grades seven and eight). \cr
   YEAR \tab \verb{  } \tab Year of data. \cr
}


Further questions regarding one-year dropout data can be sent by e-mail to
the Data Reporting Office at \url{dro@cde.ca.gov}.
}
\seealso{
\link{ca_schools}
}
\keyword{datasets}
