% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_dropouts.R
\docType{data}
\name{ca_income_tax}
\alias{ca_income_tax}
\title{Personal Income Tax Statistics By Zip Code}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 68588 rows and 8 columns.
}
\source{
California Franchise Tax Board

This data was downloaded 2023/07/27 at 13:15:00 EST and gathered from:
\url{https://data.ca.gov/dataset/personal-income-tax-statistics-by-zip-code}.
}
\usage{
ca_income_tax
}
\description{
This dataset contains data from California resident tax returns filed with
California adjusted gross income and self-assessed tax listed by zip code.
This dataset contains data for taxable years 1992 to the most recent tax
year available.
}
\details{
\tabular{lll}{
   \strong{Field Name} \tab  \tab \strong{Description} \cr
   Taxable Year \tab \verb{  } \tab Tax year in which return was filed. \cr
   Zip Code \tab \verb{  } \tab Five-digit Zip Code associated with the tax return. \cr
   State \tab \verb{  } \tab State, abbreviated, in which the return was filed. \cr
   City \tab \verb{  } \tab City name in which the return was filed. \cr
   County \tab \verb{  } \tab County name in which the return was filed. \cr
   Returns \tab \verb{  } \tab The number of returns filed for the designated taxable year and zip code. \cr
   CA AGI \tab \verb{  } \tab California adjusted gross income. \cr
   Total Tax Liability \tab \verb{  } \tab Amount of taxes owed. \cr
}


Further questions regarding this data can be sent by e-mail to
FTB Open Data at \url{no-reply@data.ftb.ca.gov}.
}
\seealso{
\link{ca_schools},\link{ca_dropouts}
}
\keyword{datasets}
