% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_dropouts.R
\docType{data}
\name{ca_schools}
\alias{ca_schools}
\title{California Public Schools and Districts}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18288 rows and 47 columns.
}
\source{
California Department of Education Data Reporting Office

This data was downloaded Mon Jan 23 11:37:26 2023 and gathered from:
\url{https://www.cde.ca.gov/ds/si/ds/pubschls.asp}.
}
\usage{
ca_schools
}
\description{
Contains all active, pending, closed, and merged public schools and
districts.
}
\details{
\tabular{lcl}{
   \strong{Field Name} \tab  \tab \strong{Description} \cr
   CDSCode \tab \verb{  } \tab This 14-digit code is the official, unique identification of  a  school within California. The first two digits identify the county, the  next  five digits identify the school district, and the last seven  digits identify  the school. Please note that a CDS code ending in '0000000'  indicates a district record not a school. \cr
   NCESDist \tab \verb{  } \tab This field represents the 7-digit National Center for Educational  Statistics (NCES)  school district identification number. The first 2  digits identify the state  and the last 5 digits identify the school  district. Combined, they make a  unique 7-digit ID for each school  district. \cr
   NCESSchool \tab \verb{  } \tab This field represents the 5-digit NCES school identification number.  The NCESSchool combined with the NCESDist form a unique 12-digit ID for  each school. \cr
   StatusType \tab \verb{  } \tab This field identifies the status of the district. Definitions of the valid status types are listed below: \cr
    \tab \verb{  } \tab ∙ \emph{Active:} The district is in operation and providing instructional services. \cr
    \tab \verb{  } \tab ∙ \emph{Closed:} The district is not in operation and no longer providing instructional services. \cr
    \tab \verb{  } \tab ∙ \emph{Merged:} The district has combined with another district or districts. \cr
    \tab \verb{  } \tab ∙ \emph{Pending:} The district has not opened for operation and instructional services yet, but plans to open within the next 9–12 months. \cr
   County \tab \verb{  } \tab County name. \cr
   District \tab \verb{  } \tab District or Administrative Authority name. \cr
   School \tab \verb{  } \tab School name. \cr
   Street \tab \verb{  } \tab The unabbreviated street (physical) address of this school, district, or administrative authority. Note: Some entities (primarily closed or retired schools) may not have data in this field. \cr
   StreetAbr \tab \verb{  } \tab The abbreviated street address of the school, district, or  administrative  authority’s physical location. Note: Some records  (primarily records of closed  or retired schools) may not have data in  this field. \cr
   City \tab \verb{  } \tab The city where the school, district, or administrative authority  is physically  located. Note: Some records (primarily records of closed  or retired schools)  may not have data in this field. \cr
   Zip \tab \verb{  } \tab The zip code associated with the street address of the school,  district, or  administrative authority’s physical location. Note: Some  records (primarily  records of closed or retired schools) may not have  data in this field. \cr
   State \tab \verb{  } \tab The state in which the school, district, or administrative authority is located. \cr
   MailStreet \tab \verb{  } \tab The unabbreviated mailing address of the school, district, or  administrative  authority (i.e., P.O. Box). Note: 1) Some entities  (primarily closed or retired  schools) may not have data in this field;  2) Many active entities have not  provided a mailing street address. For  your convenience we have filled the  unpopulated MailStreet cells with  Street data. \cr
   MailStrAbr \tab \verb{  } \tab The abbreviated mailing street address of the school, district, or  administrative  authority (i.e., P.O. Box).Note: Many active entities  have not provided a  mailing street address. For your convenience we  have filled the unpopulated  MailStrAbr cells with StreetAbr data. \cr
   MailCity \tab \verb{  } \tab The city associated with the mailing address of the school,  district, or  administrative authority. Note: Many entities have not  provided a mailing  address city. For your convenience we have filled  the unpopulated MailCity  cells with City data. \cr
   MailZip \tab \verb{  } \tab The zip code associated with the mailing address of the school,  district, or  administrative authority. Note: Many entities have not  provided a mailing  address zip code. For your convenience we have  filled the unpopulated MailZip  cells with Zip data. \cr
   MailState \tab \verb{  } \tab The state within the mailing address. For your convenience we have filled the  unpopulated MailState cells with State data. \cr
   Phone \tab \verb{  } \tab The area code and phone number of the school, district, or administrative  authority. \cr
   Ext \tab \verb{  } \tab The phone number extension of the school, district, or administrative authority. \cr
   FaxNumber \tab \verb{  } \tab The fax number of the school, district, or administrative authority. \cr
   Email \tab \verb{  } \tab The email address of the school, district, or administrative authority. \cr
   Website \tab \verb{  } \tab The website address of the school, district, or administrative authority. \cr
   OpenDate \tab \verb{  } \tab The date the school opened. \cr
   ClosedDate \tab \verb{  } \tab The date the school closed. \cr
   Charter \tab \verb{  } \tab This field identifies a charter school. The field is coded as follows: \cr
    \tab \verb{  } \tab ∙ Y = The school is a charter \cr
    \tab \verb{  } \tab ∙ N = The school is not a charter \cr
   CharterNum \tab \verb{  } \tab The charter school number, which is a 4-digit number assigned to a charter school. \cr
   FundingType \tab \verb{  } \tab Indicates the charter school funding type. Values are as follows: \cr
    \tab \verb{  } \tab ∙ Not in CS funding model \cr
    \tab \verb{  } \tab ∙ Locally funded \cr
    \tab \verb{  } \tab ∙ Directly funded \cr
   DOC \tab \verb{  } \tab The District Ownership Code (DOC) is the numeric code used to identify the category of the Administrative Authority. \cr
    \tab \verb{  } \tab ∙ 00 – County Office of Education \cr
    \tab \verb{  } \tab ∙ 02 – State Board of Education \cr
    \tab \verb{  } \tab ∙ 03 – Statewide Benefit Charter \cr
    \tab \verb{  } \tab ∙ 31 – State Special Schools \cr
    \tab \verb{  } \tab ∙ 34 – Non-school Location* \cr
    \tab \verb{  } \tab ∙ 52 – Elementary School District \cr
    \tab \verb{  } \tab ∙ 54 – Unified School District \cr
    \tab \verb{  } \tab ∙ 56 – High School District \cr
    \tab \verb{  } \tab ∙ 98 – Regional Occupational Center/Program (ROC/P) \cr
    \tab \verb{  } \tab *Only the California Education Authority has been included in the non-school location category. \cr
   DOCType \tab \verb{  } \tab The District Ownership Code Type is the text description of the DOC category. (See  text values in DOC field description above) \cr
   SOC \tab \verb{  } \tab The School Ownership Code is a numeric code used to identify the type of school. \cr
    \tab \verb{  } \tab ∙ 08 – Preschool \cr
    \tab \verb{  } \tab ∙ 09 – Special Education Schools (Public) \cr
    \tab \verb{  } \tab ∙ 10 – County Community \cr
    \tab \verb{  } \tab ∙ 11 – Youth Authority Facilities (CEA) \cr
    \tab \verb{  } \tab ∙ 13 – Opportunity Schools \cr
    \tab \verb{  } \tab ∙ 14 – Juvenile Court Schools \cr
    \tab \verb{  } \tab ∙ 15 – Other County or District Programs \cr
    \tab \verb{  } \tab ∙ 31 – State Special Schools \cr
    \tab \verb{  } \tab ∙ 60 – Elementary School (Public) \cr
    \tab \verb{  } \tab ∙ 61 – Elementary School in 1 School District (Public) \cr
    \tab \verb{  } \tab ∙ 62 – Intermediate/Middle Schools (Public) \cr
    \tab \verb{  } \tab ∙ 63 – Alternative Schools of Choice \cr
    \tab \verb{  } \tab ∙ 64 – Junior High Schools (Public) \cr
    \tab \verb{  } \tab ∙ 65 – K-12 Schools (Public) \cr
    \tab \verb{  } \tab ∙ 66 – High Schools (Public) \cr
    \tab \verb{  } \tab ∙ 67 – High Schools in 1 School District (Public) \cr
    \tab \verb{  } \tab ∙ 68 – Continuation High Schools \cr
    \tab \verb{  } \tab ∙ 69 – District Community Day Schools \cr
    \tab \verb{  } \tab ∙ 70 – Adult Education Centers \cr
    \tab \verb{  } \tab ∙ 98 – Regional Occupational Center/Program (ROC/P) \cr
   SOCType \tab \verb{  } \tab The School Ownership Code Type is the text description of  the type of school. (See text values in SOC field description above) \cr
   EdOpsCode \tab \verb{  } \tab The Education Option Code is a short text description of the type of education offered. \cr
    \tab \verb{  } \tab ∙ ALTSOC – Alternative School of Choice \cr
    \tab \verb{  } \tab ∙ COMM – County Community School \cr
    \tab \verb{  } \tab ∙ COMMDAY – Community Day School \cr
    \tab \verb{  } \tab ∙ CON – Continuation School \cr
    \tab \verb{  } \tab ∙ JUV – Juvenile Court School \cr
    \tab \verb{  } \tab ∙ OPP – Opportunity School \cr
    \tab \verb{  } \tab ∙ YTH – Youth Authority School \cr
    \tab \verb{  } \tab ∙ SSS – State Special School \cr
    \tab \verb{  } \tab ∙ SPEC – Special Education School \cr
    \tab \verb{  } \tab ∙ TRAD – Traditional \cr
    \tab \verb{  } \tab ∙ ROP – Regional Occupational Program \cr
    \tab \verb{  } \tab ∙ HOMHOS – Home and Hospital \cr
    \tab \verb{  } \tab ∙ SPECON – District Consortia Special Education School \cr
   EdOpsName \tab \verb{  } \tab The Educational Option Name is the long text description of  the type of education being offered. (See text values in EdOPsCode field description above) \cr
   EILCode \tab \verb{  } \tab The Educational Instruction Level Code is a short text description of the institution's type relative to the grade range served. \cr
    \tab \verb{  } \tab ∙ A – Adult \cr
    \tab \verb{  } \tab ∙ ELEM – Elementary \cr
    \tab \verb{  } \tab ∙ ELEMHIGH – Elementary-High Combination \cr
    \tab \verb{  } \tab ∙ HS – High School \cr
    \tab \verb{  } \tab ∙ INTMIDJR – Intermediate/Middle/Junior High \cr
    \tab \verb{  } \tab ∙ PS – Preschool \cr
    \tab \verb{  } \tab ∙ UG – Ungraded \cr
   EILName \tab \verb{  } \tab The Educational Instruction Level Name is the long text description of the institution’s type relative to the grade range  served. (See text values in EILCode field description  above) \cr
   GSoffered \tab \verb{  } \tab The grade span offered is the lowest grade and the highest grade  offered or supported by the school, district, or administrative   authority. This field might differ from the grade span served as  reported in  the most recent certified California Longitudinal Pupil  Achievement (CALPADS)  Fall 1 data collection. For example XYZ School might display the following data: GSoffered = P–Adult, GSserved = K–12 \cr
   GSserved \tab \verb{  } \tab This field represents the grade span served. It is the  lowest grade and the highest grade of student enrollment as reported in the   most recent certified CALPADS Fall 1 data collection. Only K–12  enrollment is  reported through CALPADS. This field may differ from the  grade span offered. \cr
    \tab \verb{  } \tab ∙ Note: Special programs at independent study, alternative education,  and special  education schools will often exceed the typical grade span  for schools of that  type (i.e., a high school might serve grades three  through twelve). \cr
   Virtual \tab \verb{  } \tab This field identifies the type of virtual instruction offered by  the school. Virtual instruction is instruction in which students and  teachers are separated by time and/or location, and interaction occurs  via computers and/or telecommunications technologies. The field is coded as follows: \cr
    \tab \verb{  } \tab ∙ F = Exclusively Virtual – The school has no physical building where students meet with each other or with teachers, all instruction is virtual. \cr
    \tab \verb{  } \tab ∙ V = Primarily Virtual – The school focuses on a systematic program of virtual instruction but includes some physical meetings among students or with teachers. \cr
    \tab \verb{  } \tab ∙ C = Primarily Classroom – The school offers virtual courses but virtual instruction is not the primary means of instruction. \cr
    \tab \verb{  } \tab ∙ N = Not Virtual – The school does not offer any virtual instruction. \cr
    \tab \verb{  } \tab ∙ P = Partial Virtual – The school offers some, but not all, instruction through virtual instruction. Note: This value was retired and replaced with the Primarily Virtual and Primarily Classroom values beginning with the 2016–17 school year. \cr
   Magnet \tab \verb{  } \tab This field identifies whether a school is a magnet school  and/or provides a magnet program. Additional information on magnet  schools may be obtained from the CDE Magnets web page. The field is coded as follows: \cr
    \tab \verb{  } \tab ∙ Y = Magnet - The school is a magnet school and/or offers a magnet program. \cr
    \tab \verb{  } \tab ∙ N = Not Magnet - The school is not a magnet school and/or does not offer a magnet program. \cr
    \tab \verb{  } \tab ∙ Note: Preschools and adult education centers do not contain a magnet school indicator. \cr
   YearRoundYN \tab \verb{  } \tab This field indicates if the school is operating on a year round  calendar. More information about year-round education can be found by  visiting the CDE Year-Round Education Program Guide web page. \cr
   FederalDFCDistrictID \tab \verb{  } \tab This field represents the 7-digit National Center for Educational Statistics (NCES) school district identification number for direct funded  charter schools. Beginning in February 2019, direct funded charter schools were  issued their own federal district ID and should nolonger use the federal  district ID associated with their authorizing agency. \cr
   Latitude \tab \verb{  } \tab The angular distance (expressed in degrees) between the location  of the school,  district, or administrative authority and the equator  measured north to south. \cr
   Longitude \tab \verb{  } \tab The angular distance (expressed in degrees) between the location  of the school,  district, or administrative authority and the prime  meridian (Greenwich,  England) measured from west to east. \cr
   AdmFName \tab \verb{  } \tab The superintendent’s or principal’s first name. Only active and  pending districts  and schools will display administrator information,  if applicable. \cr
   AdmLName \tab \verb{  } \tab The superintendent’s or principal’s last name. Only active and  pending districts  and schools will display administrator information,  if applicable. \cr
   AdmEmail \tab \verb{  } \tab The superintendent’s or principal’s e-mail address. Only active  and pending  districts and schools will display administrator  information, if applicable. \cr
   LastUpDate \tab \verb{  } \tab The date the record was last updated. \cr
}


Further questions regarding this data can be sent by e-mail to
the Data Reporting Office at \url{dro@cde.ca.gov}.
}
\seealso{
\link{ca_dropouts},\link{ca_income_tax}
}
\keyword{datasets}
