% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplay-package.R
\docType{package}
\name{dataplay-package}
\alias{dataplay}
\alias{dataplay-package}
\title{dataplay: Load data to play with}
\description{
This is a super simple package meant to load selected data into R for the purposes of teaching and learning.
}
\author{
\strong{Maintainer}: Ava Hoffman \email{avamariehoffman@gmail.com} (\href{https://orcid.org/0000-0002-1833-4397}{ORCID})

}
\keyword{internal}
