% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_file.R
\name{in_file}
\alias{in_file}
\title{Read Data File}
\usage{
in_file(...)
}
\arguments{
\item{...}{Arguments to pass to
\code{\link{system.file}}}
}
\value{
A character vector
}
\description{
Read Data File
}
\examples{
in_file('nsf_survey_of_earned_doctorates.csv')
}
