% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplay-package.R
\docType{package}
\name{dataplay-package}
\alias{dataplay}
\alias{dataplay-package}
\title{\code{dataplay}: Load data to play with}
\description{
This is a super simple package meant to load selected data into R for the purposes of teaching and learning.
}
\details{
Here is a list of the datasets. They are already stored as objects when you
load the package.
\itemize{
\item \code{ca_dropouts}: Historical data for grade seven through twelve dropouts and
enrollment by school in California. Data is further subdivided by
race/ethnic designation and gender.
\item \code{ca_schools}: Contains all active, pending, closed, and merged public
schools and districts in California.
\item \code{doctorate_survey}: The Survey of Earned Doctorates (SED), an annual
census of individuals who earn research doctoral degrees from accredited
U.S. academic institutions.
}
}
\author{
\strong{Maintainer}: Ava Hoffman \email{avamariehoffman@gmail.com} (\href{https://orcid.org/0000-0002-1833-4397}{ORCID})

}
\keyword{internal}
