% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earned_doctorates.R
\name{doctorate_survey}
\alias{doctorate_survey}
\title{Read NSF Survey of earned doctorates data}
\usage{
doctorate_survey(...)
}
\arguments{
\item{...}{Additional arguments to pass to
\code{\link{read.csv}}}
}
\value{
A \code{tibble}
}
\description{
Read NSF Survey of earned doctorates data
}
\examples{
df <- doctorate_survey()
}
